프로필이미지 변경가능

+++ WebContent/profimgAction.jsp
@@ -33,25 +33,35 @@ String encType="utf-8"; // 변환 방식
 int maxSize = 5*1024*1024; // 사진의 size
 
 ServletContext context = this.getServletContext(); //절대경로를 얻는다.
-String realFolder = context.getRealPath("image"); //saveFolder의 절대경로를 받는다.
+String realFolder = context.getRealPath("image"); //image폴더의 절대경로를 받는다.
 System.out.println(realFolder);
 MultipartRequest multi = null;
-
 
 //파일업로드를 실질적으로 담당하는 부분
 multi = new MultipartRequest(request,realFolder,maxSize,encType,new DefaultFileRenamePolicy());	
 
+//form으로 전달받은 파일이름 가져온다.
+String fileName = multi.getFilesystemName("fileName"); 
 
-//form으로 전달받은 파일이름 가져온다.\
-String fileName = multi.getFilesystemName("fileName"); //파일이름
+//이전 사진명
+File beforeImg = new File(realFolder+"\\"+studentNum+"프로필사진.jpg") ;
 
 
 if(fileName != null) {
-	File oldFile = new File(realFolder+"\\"+fileName);
-	File newFile = new File(realFolder+"\\"+studentNum+"프로필사진.jpg");
-	oldFile.renameTo(newFile);
-	System.out.println("저장된 파일 이름: "+newFile);
-	
+	if(beforeImg.exists()) { //이전 프로필사진 있으면
+		beforeImg.delete(); //이전 사진 삭제
+		//사진업로드
+		File oldFile = new File(realFolder+"\\"+fileName);
+		File newFile = new File(realFolder+"\\"+studentNum+"프로필사진.jpg");
+		oldFile.renameTo(newFile);
+		System.out.println("저장된 파일 이름: "+newFile);
+		
+	} else { //사진업로드
+		File oldFile = new File(realFolder+"\\"+fileName);
+		File newFile = new File(realFolder+"\\"+studentNum+"프로필사진.jpg");
+		oldFile.renameTo(newFile);
+		System.out.println("저장된 파일 이름: "+newFile);
+	}
 	
 }
